# swagger.yaml
openapi: 3.0.0
info:
  title: Sabr Shukr Backend API
  version: 1.0.0
  description: API documentation for the Sabr Shukr e-commerce platform backend.
servers:
  - url: http://localhost:5000/api
    description: Development server
  - url: https://sabrshukrbackend.onrender.com/api
    description: Production server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    RegisterWithReferralRequest:
      type: object
      required:
        - email
        - password
        - fullName
      properties:
        email:
          type: string
          format: email
          example: user@example.com
        password:
          type: string
          format: password
          example: strongpassword123
        fullName:
          type: string
          example: John Doe
        referralCode:
          type: string
          description: Optional referral code used during registration.
          example: "FREESHIP"
    otpLoginRequest:
      type: object
      required:
        - phone
        - otp
      properties:
        phone:
          type: string
          description: Phone number in E.164 format (e.g., +919876543210)
          example: "+919876543210"
        otp:
          type: string
          description: The OTP code sent to the user's phone
          example: "123456"
        fullName:
          type: string
          description: Optional full name to store in profiles table
          example: "John Doe"
        email:
          type: string
          format: email
          description: Optional email to update in auth.users
          example: "user@example.com"
    sendOtpRequest:
      type: object
      required:
        - phone
      properties:
        phone:
          type: string
          description: Phone number in E.164 format (e.g., +919876543210)
          example: "+919876543210"
...
paths:
  /auth/register:
    post:
      tags:
        - Auth
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterWithReferralRequest'
      responses:
        '201':
          description: User successfully registered
        '400':
          description: Invalid input
  /auth/login:
    post:
      tags:
        - Auth
      summary: Log in an existing user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                  example: user@example.com
                password:
                  type: string
                  format: password
                  example: strongpassword123
      responses:
        '200':
          description: User successfully logged in
        '401':
          description: Invalid credentials
  /auth/logout:
    post:
      tags:
        - Auth
      summary: Logs out the currently authenticated user
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Successfully logged out
        '401':
          description: Unauthorized
  /auth/verify-token:
    get:
      tags:
        - Auth
      summary: Verifies if the provided Bearer token is valid
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Token is valid
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    description: Supabase auth user payload
        '401':
          description: Invalid or expired token / No token provided
        '500':
          description: Internal Server Error
  /auth/social-login:
    post:
      tags:
        - Auth
      summary: Initiates a social login flow
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                provider:
                  type: string
                  enum: [google, facebook]
                  example: google
      responses:
        '200':
          description: Social login initiated. Returns a redirect URL.
        '400':
          description: Invalid provider
  /auth/refresh-token:
    post:
      tags:
        - Auth
      summary: Refreshes an access token using a valid refresh token
      parameters:
        - in: header
          name: X-Refresh-Token
          required: false
          schema:
            type: string
          description: Optional. Provide refresh token via header instead of body. If provided, body is not required.
        - in: header
          name: Authorization
          required: false
          schema:
            type: string
            example: Bearer <refreshToken>
          description: Optional. You can also pass the refresh token as a Bearer token in the Authorization header for this endpoint.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - refreshToken
              properties:
                refreshToken:
                  type: string
                  example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
      responses:
        '200':
          description: Successfully refreshed session
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    description: The new session and tokens payload from Supabase
        '400':
          description: Missing refreshToken
        '401':
          description: Invalid or expired refresh token
        '500':
          description: Internal Server Error
...
  /auth/check-phone-exists:
    post:
      tags:
        - Auth
      summary: Checks if a phone number is already registered in Supabase auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - phone
              properties:
                phone:
                  type: string
                  example: "+919876543210"
      responses:
        '200':
          description: Phone existence check result
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  exists:
                    type: boolean
        '400':
          description: Missing or invalid input
        '500':
          description: Internal Server Error
    put:
      tags:
        - User Profile & Addresses
      summary: Updates the profile of the currently logged-in user
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                fullName:
                  type: string
                  example: Jane Doe
                profilePictureUrl:
                  type: string
                  format: uri
                  example: https://example.com/profile.jpg
                dateOfBirth:
                  type: string
                  format: date
                  example: 1990-05-21
      responses:
        '200':
          description: User profile updated successfully
        '400':
          description: Invalid input
        '401':
          description: Unauthorized
        '404':
          description: User profile not found
        '500':
          description: Internal Server Error
...
  /auth/loginwithOtp:
    post:
      tags:
        - Auth
      summary: Log in a user using phone number and OTP
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/otpLoginRequest'
      responses:
        '200':
          description: User successfully logged in with OTP
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    description: Supabase session/user data
        '400':
          description: Invalid input or OTP
        '401':
          description: Invalid credentials or OTP
        '500':
          description: Internal Server Error
  /auth/sendOtp:
    post:
      tags:
        - Auth
      summary: Send an OTP to a user's phone number for login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/sendOtpRequest'
      responses:
        '200':
          description: OTP sent successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: OTP sent successfully
                  data:
                    type: object
                    description: Supabase response data
        '400':
          description: Invalid input or phone number
        '500':
          description: Internal Server Error
...